webwatcher/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pytesseract==0.3.13
Pillow==10.4.0
Jinja2==3.1.4
APScheduler==3.10.4
httpx==0.27.2
playwright==1.47.0